mostly done, single player works
