+++ Assets/Bundles/Prefabs/Collectables/WorldItems/CollectableKey0.prefab
@@ -1,0 +1,139 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!1001 &6329307037579920239
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 1160627986272127395, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: NetworkGuid.RawGuidValue.Array.data[0]
+      value: -6557278183554499752
+      objectReference: {fileID: 0}
+    - target: {fileID: 1160627986272127395, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: NetworkGuid.RawGuidValue.Array.data[1]
+      value: 6756529056473328506
+      objectReference: {fileID: 0}
+    - target: {fileID: 3676629508602288210, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: _Collected._items.Array.size
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4869834720623374894, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_Materials.Array.size
+      value: 3
+      objectReference: {fileID: 0}
+    - target: {fileID: 4869834720623374894, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_Materials.Array.data[0]
+      value: 
+      objectReference: {fileID: 2100000, guid: c95ad62c7efe6b641a2cdba628d6c8ec, type: 2}
+    - target: {fileID: 4869834720623374894, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_Materials.Array.data[1]
+      value: 
+      objectReference: {fileID: 2100000, guid: 54b7d83ec25ce214a8369457c4d11b8e, type: 2}
+    - target: {fileID: 4869834720623374894, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_Materials.Array.data[2]
+      value: 
+      objectReference: {fileID: 2100000, guid: 0f013e4c55ed7b7419c678291fff04d7, type: 2}
+    - target: {fileID: 4869834720623374894, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_Materials.Array.data[3]
+      value: 
+      objectReference: {fileID: 2100000, guid: 0f013e4c55ed7b7419c678291fff04d7, type: 2}
+    - target: {fileID: 4869834720623374894, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_Materials.Array.data[4]
+      value: 
+      objectReference: {fileID: 2100000, guid: 103b71d14fa728f45a531a187b6efedc, type: 2}
+    - target: {fileID: 5915339226451451560, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_RootOrder
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5915339226451451560, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5915339226451451560, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5915339226451451560, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5915339226451451560, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 5915339226451451560, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5915339226451451560, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5915339226451451560, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5915339226451451560, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5915339226451451560, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5915339226451451560, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6416654292541759145, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: collectableType
+      value: 10
+      objectReference: {fileID: 0}
+    - target: {fileID: 6439110641031758866, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_Name
+      value: CollectableKey0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6514375072736587760, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_LocalScale.x
+      value: 0.3
+      objectReference: {fileID: 0}
+    - target: {fileID: 6514375072736587760, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_LocalScale.y
+      value: 0.3
+      objectReference: {fileID: 0}
+    - target: {fileID: 6514375072736587760, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_LocalScale.z
+      value: 0.2
+      objectReference: {fileID: 0}
+    - target: {fileID: 6514375072736587760, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6514375072736587760, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6514375072736587760, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: -0.095
+      objectReference: {fileID: 0}
+    - target: {fileID: 6514375072736587760, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6514375072736587760, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6514375072736587760, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_ConstrainProportionsScale
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8264275471743424186, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
+      propertyPath: m_Mesh
+      value: 
+      objectReference: {fileID: 4300000, guid: 1c3e4b442bd1c7c46a315c4f32321b80, type: 2}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: e37400ca93ab33e48aee39750ee1eec6, type: 3}
