+++ Assets/Bundles/Prefabs/Collectables/WorldItems/CollectableKey0.prefab.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 58bbd0e73bdeffa47aabb2156103c45d
+labels:
+- FusionPrefab
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
