+++ Assets/Bundles/Prefabs/Islands/маяк.prefab.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 880e20aa123286942ad0a8d2582f3864
+labels:
+- FusionPrefab
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
