+++ Assets/Models/Items/key.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 31d0c8cf7ca2fc84597f2183c092140f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
