+++ Assets/Models/Items/key/key.mesh.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1c3e4b442bd1c7c46a315c4f32321b80
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 4300000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
