+++ Assets/Models/Items/key/белый.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2d70294a62c14444b9f60af937683bf6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
