+++ Assets/Models/Items/key/зелёный.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0f013e4c55ed7b7419c678291fff04d7
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
