+++ Assets/Models/Items/key/красный.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 103b71d14fa728f45a531a187b6efedc
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
