+++ Assets/Models/Items/key/синий.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 54b7d83ec25ce214a8369457c4d11b8e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
