+++ Assets/Models/Items/key/чёрный.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c95ad62c7efe6b641a2cdba628d6c8ec
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
